Files to help with debugging
The csv file is where the data is being extracted by the beginning of the script. The .TXT file is to show an example results file of inputting "2.0C-H4A-DC2".
When the entire script runs properly there should be a second .txt file with the word "_filter" attached listing results of search without copies, which you can see clearly if you compare some lines from the file.

I believe the code to find a duplicate is wrong. 
One result format is this : "2320  Logical ID:2320 192.168.203.216" 
and the other is just "2320  192.168.203.216"  without "Logical ID:2320"

The program is not seeing those two lines as copies of each other because they are not exact.